Reduce sleep time to 4 seconds
Seems going lower results in getting blocked. Without any sleep, the gap between requests is about 5 or 6 seconds on average so the sleep would push the gap to about 10 seconds. I suspect that sub 10 seconds is where the anti-scraping measures kick in. Could be sub 5 seconds as well since I noticed some events saved a bit faster than others. But not going to fiddle too much with this value now. Seems like it's at a value where we're almost guaranteed to not get blocked but also is reasonably quick.